requirements.txt
numpy           >=  1.19.5
torch           ==  1.7.0
torchvision     ==  0.8.1
opencv-python   >=  4.5.1.48
redis           >=  3.5.3
tqdm            >=  4.60.0
pymysql         >=  1.0.2
pillow          >=  8.2.0
Flask           >=  1.1.2
easydict        >=  1.9
matplotlib      >=  3.3.4
Shapely         >=  1.7.1
xxhash          >=  2.0.0
werkzeug        >=  1.0.1
six             >=  1.15.0
scipy           >=  1.5.4
traitlets       >= 4.3.3
pyyaml          >= 5.4.1
# moviepy         ==  1.0.3
tensorflow      ==  2.1.0
tensorflow-hub  ==  0.7.0

## Notice ##
# python version must least 3.7

